component: Dimensions - halfWidth/halfHeight are now uint32_t instead of float

+++ spn/include/component/Dimensions.h
@@ -13,7 +13,7 @@ namespace component
 
 	public:
 		uint32_t width, height;
-		float halfWidth, halfHeight;
+		uint32_t halfWidth, halfHeight;
 	};
 }
